[INFINITY-3202] Remove partition tests, save one. (#2300)
* Remove recovery_partition suite
Move the one "useful" partition test into test_zzzrecovery

* Move test to be before the "I MUST BE LAST" test.
